+++ public/index.html
@@ -41,6 +41,6 @@
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
     <div id="feedback-modal-root"></div>
-    <!-- <div id="task-modal-root"></div> -->
+    <div id="task-modal-root"></div>
   </body>
 </html>
