+++ src/components/AuthorizedUserComponents/Header/HeaderItems/AddFeedbackBtn.jsx
@@ -2,16 +2,22 @@ import React, { useState, useEffect } from 'react';
 import { Button } from './AddFeedbackBtn.styled';
 import AddFeedbackModal from '../../AddFeedbackModal/AddFeedbackModal';
 
+// import TaskModal from 'components/CommonComponents/TaskModal/TaskModal';
+
+
 const AddFeedbackBtn = () => {
   const [showModal, setShowModal] = useState(false);
   const [isModalOpen, setIsModalOpen] = useState(false);
 
   /// Toggle Modal Function ///
   const toggleModal = () => {
+    console.log('Toggle modal called');
+    console.log('Show modal before:', showModal);
     setShowModal(!showModal);
+   console.log('Show modal after:', showModal);
   };
 
-/// Not to scroll page when open modal ///
+/// Not to scroll page when modal open ///
   useEffect(() => {
     if (showModal) {
       setIsModalOpen(true);
@@ -28,6 +34,7 @@ const AddFeedbackBtn = () => {
         Feedback
       </Button>
       {isModalOpen && <AddFeedbackModal toggleModal={toggleModal} />}
+      {/* {isModalOpen && <TaskModal toggleModal={toggleModal} />} */}
     </>
   );
 };
