+++ src/components/CommonComponents/TaskModal/TaskModal.styled.jsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const CloseModalButton = styled.button`
+  position: absolute;
+  right: 15px;
+  top: 15px;
+  border: none;
+  background-color: transparent;
+`;
+
+export { CloseModalButton };
