+++ src/components/CommonComponents/TaskModal/TaskModalItems/Modal.jsx
@@ -1,5 +1,7 @@
-const Modal = () => {
-  return <div>Modal</div>;
+import { ModalWindow } from './Modal.styled';
+
+const Modal = ({ children }) => {
+  return <ModalWindow>{children}</ModalWindow>;
 };
 
 export default Modal;
