+++ src/components/GeneralComponents/ModalButton/ModalButton.jsx
@@ -1,7 +1,16 @@
 import { Button } from './ModalButton.styled';
 
 const ModalButton = ({ textColor, backgroundColor, children, ...rest }) => {
-  return <Button type='submit' textColor={textColor} backgroundColor={backgroundColor} {...rest}>{children}</Button>;
+  return (
+    <Button
+      type="submit"
+      textColor={textColor}
+      backgroundColor={backgroundColor}
+      {...rest}
+    >
+      {children}
+    </Button>
+  );
 };
 
 export default ModalButton;
