+++ src/components/GeneralComponents/ModalButton/ModalButton.styled.jsx
@@ -11,8 +11,8 @@ const Button = styled.button`
   font-weight: 14px;
   line-height: 1.29;
   font-weight: 600;
-  font-size:12px;
-  line-height:1.17;
+  font-size: 12px;
+  line-height: 1.17;
   color: ${props => props.textColor || '#343434'};
   background-color: ${props => props.backgroundColor || '#E5EDFA'};
   :not(:last-child) {
@@ -21,8 +21,8 @@ const Button = styled.button`
   @media (min-width: 768px) {
     padding-top: 15px;
     padding-bottom: 15px;
-    font-size:14px;
-  line-height:1.29;
+    font-size: 14px;
+    line-height: 1.29;
   }
 `;
