+++ src/components/GeneralComponents/ModalWindow/ModalWindow.jsx
@@ -1,7 +1,7 @@
 import { ModalWindow } from './ModalWindow.styled';
 
 const Modal = ({ children }) => {
-    return <ModalWindow>{ children}</ModalWindow>
+  return <ModalWindow>{children}</ModalWindow>;
 };
 
 export default Modal;
